% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{crab_sample}
\alias{crab_sample}
\title{Generate a sample of crabs}
\usage{
crab_sample(file)
}
\arguments{
\item{file}{A YAML file, see examples in \code{inst/samples}.}
}
\value{
A \link[tibble:tibble-package]{tibble} of fiddler crabs.
}
\description{
Takes a configuration (YAML) file as input and generates a sample of fiddler
crabs according to the specification in file.
}
