% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiddlers-package.R
\docType{package}
\name{fiddlers-package}
\alias{fiddlers}
\alias{fiddlers-package}
\title{fiddlers: A Fictional Data Set About Fiddler Crabs in the Algarve}
\description{
A fictional data set about fiddler crabs in the Algarve, Portugal. This R data package provides simple demographics and a few morphometric variables about three fictitious species of fiddler crabs. It can be useful to teach marine biologists the fundamental principles of data science, such as tidying and visualisation.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Isabel Duarte \email{iduarte.scientist@gmail.com} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
}

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
