% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fiddlers}
\alias{fiddlers}
\title{Fiddler crabs data set}
\format{
A tibble of 12 columns:
\describe{
\item{\code{species}}{Crab species name.}
\item{\code{region}}{Region in the Algarve, Portugal, where the crabs have been sampled.}
\item{\code{season}}{Season: either \code{"summer"} or \code{"winter"}.}
\item{\code{quadrat}}{Quadrat identifier.}
\item{\code{sex}}{Crab sex.}
\item{\code{stage}}{Developmental stage: \code{"juvenile"}, \code{"pre_puberty"}, \code{"sub_adult"} or \code{"adult"}.}
\item{\code{cw}}{Carapace width (in millimetres).}
\item{\code{cl}}{Carapace length (in millimetres).}
\item{\code{fw}}{Front width (in millimetres): section of the carapace in between the eyestalks.}
\item{\code{hypertrophy}}{Side of claw hypertrophy: either \code{"none"}, \code{"left"}, \code{"right"} or \code{"both"}.}
\item{\code{lcl}}{Left claw length (in millimetres).}
\item{\code{rcl}}{Right claw length (in millimetres).}
}
}
\usage{
fiddlers
}
\description{
This is a fictitious data set of 7326 crabs.
}
\keyword{datasets}
