% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quadrats}
\alias{quadrats}
\title{A data set of quadrats}
\format{
A tibble of 3 columns:
\describe{
\item{\code{region}}{Region in the Algarve, Portugal, where the crabs have been sampled.}
\item{\code{quadrat}}{Quadrat identifier. A quadrat refers to the location and area where the sampling of crabs took place.}
\item{\code{area}}{Quadrat area in squared metres.}
}
}
\usage{
quadrats
}
\description{
This is a data set of the areas (in squared metres) of the quadrats used for
sampling the crabs that comprise the \code{fiddlers} data set. This is a
fictitious data set.
}
\keyword{datasets}
